New jbigi/jcpuid compiles for Windows.
Compiled using mingw32 in Windows XP (gcc --version returns v4.5.2).

As was done with the Linux and FreeBSD builds, jbigi was linked with gmp-4.3.2
for the 32bit cpus and 5.0.2 for the 64bit CPUs.

Since there current isn't a free version of the wrapper for x64 Windows, there
aren't any x64 versions of these dlls.