Windows: Don't install as a service by default.
With this checkin I'm trying to lessen the occurences of ticket #474:

If a user installs I2P on top of an already existing I2P installation with the service
enabled, the installer will hang. The Quit button cannot be clicked. Clicking
the X in the corner seems to roll back the installation.

Also #472: service is not removed when I2P is uninstalled